Move Runway config & Eloquent models out of the `TestCase`
